LayoutTests/http/tests/security/contentSecurityPolicy/1.1/stylehash-allowed-expected.txt

ALERT: PASS (1/3): The '#p1' element's text is green, which means the style was correctly applied.
ALERT: PASS (2/3): The '#p2' element's text is green, which means the style was correctly applied.
ALERT: PASS (3/3): The '#p3' element's text is green, which means the style was correctly applied.
This tests the result of a valid style hash. It passes if this text is green, and a "PASS" alert for p1 is fired.

This tests the result of a valid style hash. It passes if this text is green, and a "PASS" alert for p2 is fired.

This tests the result of a valid style hash. It passes if this text is green, and a "PASS" alert for p3 is fired.
